refactor(meerkat-core): split sqlToMeerkat into focused modules + exact test assertions
- Split 827-line monolith into 5 modules:
  - sql-to-meerkat.ts (orchestrator, ~230 lines)
  - helpers.ts (shared utilities)
  - extract-filters.ts (WHERE/HAVING filter extraction)
  - extract-order.ts (ORDER BY logic)
  - build-base-sql.ts (AST-based base SQL construction)
- Rewrite tests to use exact SQL match assertions instead of partial checks
- Tests now verify exact rebuilt SQL strings, filter structures, and schema shapes

work-item: KERN-1560
