+++ meerkat-core/src/sql-to-meerkat/build-base-sql.ts
@@ -1,0 +1,52 @@
+import {
+  astDeserializerQuery,
+  deserializeQuery,
+} from '../ast-deserializer/ast-deserializer';
+import {
+  ExpressionClass,
+  ExpressionType,
+  ParsedExpression,
+  SelectNode,
+} from '../types/duckdb-serialization-types';
+import { GetQueryOutput } from '../utils/duckdb-ast-parse-serialize';
+import { stripQueryLocationInPlace } from './helpers';
+
+export async function buildBaseSQL(
+  originalSql: string,
+  selectNode: SelectNode,
+  hadAggregation: boolean,
+  residualWhere: ParsedExpression | undefined,
+  getQueryOutput: GetQueryOutput
+): Promise<string> {
+  const fromOnlyNode = {
+    ...selectNode,
+    select_list: [
+      {
+        class: ExpressionClass.STAR,
+        type: ExpressionType.COLUMN_REF,
+        alias: '',
+        relation_name: '',
+        exclude_list: [],
+        replace_list: [],
+        columns: false,
+      },
+    ],
+    where_clause: hadAggregation || residualWhere ? residualWhere : undefined,
+    group_expressions: [],
+    group_sets: [],
+    having: null,
+    qualify: null,
+    modifiers: [],
+  };
+
+  stripQueryLocationInPlace(fromOnlyNode);
+
+  try {
+    const query = astDeserializerQuery({ node: fromOnlyNode } as never);
+    const rows = await getQueryOutput(query);
+    const baseSql = deserializeQuery(rows).replace(/;\s*$/, '');
+    return baseSql;
+  } catch {
+    return `SELECT * FROM (${originalSql.replace(/'/g, "''")}) AS _base`;
+  }
+}
